+++ R/db-sql.R
@@ -330,20 +330,23 @@ dbplyr_query_select <- function(con, ...) {
 }
 #' @importFrom dplyr sql_select
 #' @export
-sql_select.DBIConnection <- function(con, select, from, where = NULL,
-                                     group_by = NULL, having = NULL,
-                                     window = NULL,
+sql_select.DBIConnection <- function(con,
+                                     select,
+                                     from,
+                                     where = NULL,
+                                     group_by = NULL,
+                                     having = NULL,
                                      order_by = NULL,
                                      limit = NULL,
                                      distinct = FALSE,
                                      ...,
                                      subquery = FALSE) {
+  # TODO should add argument `window` after tidyverse/dplyr#4663
   sql_query_select(
     con, select, from,
     where = where,
     group_by = group_by,
     having = having,
-    window = window,
     order_by = order_by,
     limit = limit,
     distinct = distinct,
@@ -372,7 +375,7 @@ sql_query_join.DBIConnection <- function(con, x, y, vars, type = "inner", by = N
   x <- dbplyr_sql_subquery(con, x, name = by$x_as, lvl = lvl)
   y <- dbplyr_sql_subquery(con, y, name = by$y_as, lvl = lvl)
 
-  select <- sql_join_vars(con, vars, x_as = by$x_as, y_as = by$y_as, type = type)
+  select <- sql_rf_join_vars(con, type = type, vars, x_as = by$x_as, y_as = by$y_as)
   on <- sql_join_tbls(con, by, na_matches = na_matches)
 
   # Wrap with SELECT since callers assume a valid query is returned