+++ R/lazy-join-query.R
@@ -25,13 +25,57 @@ lazy_multi_join_query <- function(x,
     cli_abort("`vars` must have fields `name`, `table`, `var`", .internal = TRUE)
   }
   check_character(vars$name, call = call)
-  check_list(vars$table, call = call)
-  check_list(vars$var, call = call)
+  check_integer(vars$table, call = call)
+  check_character(vars$var, call = call)
 
   lazy_query(
     query_type = "multi_join",
     x = x,
     joins = joins,
+    table_names = table_names,
+    vars = vars,
+    group_vars = group_vars,
+    order_vars = order_vars,
+    frame = frame
+  )
+}
+
+#' @export
+#' @rdname sql_build
+lazy_rf_join_query <- function(x,
+                               y,
+                               type,
+                               by,
+                               table_names,
+                               vars,
+                               group_vars = op_grps(x),
+                               order_vars = op_sort(x),
+                               frame = op_frame(x),
+                               call = caller_env()) {
+  check_lazy_query(x, call = call)
+  check_lazy_query(y, call = call)
+
+  check_character(type, call = call)
+
+  if (!identical(colnames(table_names), c("name", "from"))) {
+    cli_abort("`table_names` must have fields `name`, `from`", .internal = TRUE)
+  }
+  check_character(table_names$name, call = call)
+  check_character(table_names$from, call = call)
+
+  if (!identical(colnames(vars), c("name", "x", "y"))) {
+    cli_abort("`vars` must have fields `name`, `x`, `y`", .internal = TRUE)
+  }
+  check_character(vars$name, call = call)
+  check_character(vars$x, call = call)
+  check_character(vars$y, call = call)
+
+  lazy_query(
+    query_type = "rf_join",
+    x = x,
+    y = y,
+    type = type,
+    by = by,
     table_names = table_names,
     vars = vars,
     group_vars = group_vars,
@@ -127,6 +171,10 @@ op_vars.lazy_multi_join_query <- function(op) {
   op$vars$name
 }
 #' @export
+op_vars.lazy_rf_join_query <- function(op) {
+  op$vars$name
+}
+#' @export
 op_vars.lazy_semi_join_query <- function(op) {
   op$vars$name
 }
@@ -135,61 +183,26 @@ op_vars.lazy_semi_join_query <- function(op) {
 sql_build.lazy_multi_join_query <- function(op, con, ...) {
   table_names_out <- generate_join_table_names(op$table_names)
 
-  if (length(table_names_out) > 2 || !op$joins$type %in% c("full", "right")) {
-    table_vars <- purrr::map(
-      set_names(c(list(op$x), op$joins$table), table_names_out),
-      op_vars
-    )
-
-    op$joins$table <- purrr::map(op$joins$table, ~ sql_optimise(sql_build(.x, con), con))
-    op$joins$by <- purrr::map2(
-      op$joins$by, seq_along(op$joins$by),
-      function(by, i) {
-        by$x_as <- ident(table_names_out[op$joins$by_x_table_id[[i]]])
-        by$y_as <- ident(table_names_out[i + 1L])
-        by
-      }
-    )
-
-    out <- multi_join_query(
-      x = sql_optimise(sql_build(op$x, con), con),
-      joins = op$joins,
-      table_vars = table_vars,
-      vars = op$vars
-    )
-
-    return(out)
-  }
-
-  # construct a classical `join_query()` so that special handling of
-  # full and right join continue to work
-  type <- op$joins$type
-  x_idx <- purrr::map_lgl(op$vars$table, ~ 1L %in% .x)
-  vars_x <- purrr::map2_chr(op$vars$var, x_idx, ~ {if (.y) .x[[1]] else NA_character_})
-
-  y_idx <- purrr::map_lgl(op$vars$table, ~ 2L %in% .x)
-  vars_y <- purrr::map2_chr(op$vars$var, y_idx, ~ {if (.y) dplyr::last(.x) %||% .x[[1]] else NA_character_})
-
-  vars_classic <- list(
-    alias = op$vars$name,
-    x = vars_x,
-    y = vars_y,
-    all_x = op_vars(op$x),
-    all_y = op_vars(op$joins$table[[1]])
+  table_vars <- purrr::map(
+    set_names(c(list(op$x), op$joins$table), table_names_out),
+    op_vars
   )
 
-  by <- op$joins$by[[1]]
-  by$x_as <- ident(table_names_out[[1]])
-  by$y_as <- ident(table_names_out[[2]])
+  op$joins$table <- purrr::map(op$joins$table, ~ sql_optimise(sql_build(.x, con), con))
+  op$joins$by <- purrr::map2(
+    op$joins$by, seq_along(op$joins$by),
+    function(by, i) {
+      by$x_as <- ident(table_names_out[op$joins$by_x_table_id[[i]]])
+      by$y_as <- ident(table_names_out[i + 1L])
+      by
+    }
+  )
 
-  join_query(
-    sql_optimise(sql_build(op$x, con), con),
-    sql_optimise(sql_build(op$joins$table[[1]], con), con),
-    vars = vars_classic,
-    type = type,
-    by = by,
-    suffix = NULL, # it seems like the suffix is not used for rendering
-    na_matches = by$na_matches
+  multi_join_query(
+    x = sql_optimise(sql_build(op$x, con), con),
+    joins = op$joins,
+    table_vars = table_vars,
+    vars = op$vars
   )
 }
 
@@ -220,6 +233,29 @@ generate_join_table_names <- function(table_names) {
     # don't over anchor to the start of the string,
     # since we opt to add qualifiers (...1, _{R/L}HS, etc.) to end of table name
     method = "both.sides"
+  )
+}
+
+#' @export
+sql_build.lazy_rf_join_query <- function(op, con, ...) {
+  table_names_out <- generate_join_table_names(op$table_names)
+
+  vars_classic <- as.list(op$vars)
+  vars_classic$all_x <- op_vars(op$x)
+  vars_classic$all_y <- op_vars(op$y)
+
+  by <- op$by
+  by$x_as <- ident(table_names_out[[1]])
+  by$y_as <- ident(table_names_out[[2]])
+
+  join_query(
+    sql_optimise(sql_build(op$x, con), con),
+    sql_optimise(sql_build(op$y, con), con),
+    vars = vars_classic,
+    type = op$type,
+    by = by,
+    suffix = NULL, # it seems like the suffix is not used for rendering
+    na_matches = by$na_matches
   )
 }