+++ tests/testthat/test-backend-mysql.R
@@ -34,18 +34,21 @@ test_that("custom stringr functions translated correctly", {
 # verbs -------------------------------------------------------------------
 
 test_that("generates custom sql", {
-  con <- simulate_mysql()
+  con_maria <- simulate_mysql()
 
-  expect_snapshot(sql_table_analyze(con, in_schema("schema", "tbl")))
-  expect_snapshot(sql_query_explain(con, sql("SELECT * FROM table")))
+  expect_snapshot(sql_table_analyze(con_maria, in_schema("schema", "tbl")))
+  expect_snapshot(sql_query_explain(con_maria, sql("SELECT * FROM table")))
 
-  lf <- lazy_frame(x = 1, con = con)
+  lf <- lazy_frame(x = 1, con = con_maria)
   expect_snapshot(left_join(lf, lf, by = "x", na_matches = "na"))
   expect_snapshot(error = TRUE, full_join(lf, lf, by = "x"))
 
   expect_snapshot(slice_sample(lf, n = 1))
 
-  expect_snapshot(copy_inline(con, tibble(x = 1:2, y = letters[1:2])) %>% remote_query())
+  expect_snapshot(copy_inline(con_maria, tibble(x = 1:2, y = letters[1:2])) %>% remote_query())
+
+  con_mysql <- simulate_dbi("MySQLConnection")
+  expect_snapshot(copy_inline(con_mysql, tibble(x = 1:2, y = letters[1:2])) %>% remote_query())
 })
 
 test_that("`sql_query_update_from()` is correct", {
@@ -98,9 +101,9 @@ test_that("can update", {
   con <- src_test("MariaDB")
 
   df_x <- tibble(a = 1:3, b = 11:13, c = 1:3, d = c("a", "b", "c"))
-  x <- copy_to(con, df_x, "df_x", temporary = TRUE, overwrite = TRUE)
+  x <- local_db_table(con, df_x, "df_x")
   df_y <- tibble(a = 2:3, b = c(12L, 13L), c = -(2:3), d = c("y", "z"))
-  y <- copy_to(con, df_y, "df_y", temporary = TRUE, overwrite = TRUE) %>%
+  y <- local_db_table(con, df_y, "df_y") %>%
     mutate(c = c + 1)
 
   # `RETURNING` in an `UPDATE` clause is not (yet) supported for MariaDB