+++ tests/testthat/test-verb-select.R
@@ -166,8 +166,8 @@ test_that("select() after left_join() is inlined", {
       relocate(b))
   )
   expect_equal(op_vars(out), c("b", "x", "a"))
-  expect_equal(out$lazy_query$vars$var, list("b", "x", "a"))
-  expect_equal(out$lazy_query$vars$table, list(2L, 1L, 1L))
+  expect_equal(out$lazy_query$vars$var, c("b", "x", "a"))
+  expect_equal(out$lazy_query$vars$table, c(2L, 1L, 1L))
 
   out <- left_join(lf1, lf2, by = "x") %>%
       transmute(b, x = x + 1)
@@ -227,8 +227,8 @@ test_that("select() after join handles previous select", {
 
   expect_equal(op_vars(lf2), c("x2", "y3", "z"))
   vars2 <- lf2$lazy_query$vars
-  expect_equal(vars2$var, list("x", "y", "z"))
-  expect_equal(vars2$table, list(1L, 1L, 1L))
+  expect_equal(vars2$var, c("x", "y", "z"))
+  expect_equal(vars2$table, c(1L, 1L, 1L))
 
   expect_equal(op_grps(lf2), c("x2", "y3", "z"))
   expect_snapshot(print(lf2))